seperated main.js file
