+++ assets/js/dom.js
@@ -1,10 +1,10 @@
 // eslint-disable-next-line import/extensions
 import { GameBoard, GameRound } from './app.js';
 
-const round = GameRound();
+export const round = GameRound();
 
-const displayController = (() => {
-  const renderBoard = () => {
+export const displayController = (() => {
+  const renderBoard = (initializeGame) => {
     GameBoard.testBoard.forEach((letter, index) => {
       const block = document.querySelector(`.block${index + 1}`);
       if (GameBoard.gamesPlayed === 0) {
@@ -14,7 +14,7 @@ const displayController = (() => {
     });
   };
 
-  const btnListners = () => {
+  const btnListners = (initializeGame) => {
     const playBtn = document.querySelector('#play');
     const resetBtn = document.querySelector('#reset');
     playBtn.addEventListener('click', initializeGame);
@@ -27,16 +27,16 @@ const displayController = (() => {
 })();
 
 
-const initializeGame = () => {
-  GameBoard.resetBoard();
-  if (GameBoard.gamesPlayed === 0) {
-    round.setPlayerNames();
-  }
-  round.resetPlayerArrays();
-  round.moveCount = 0;
-  displayController.renderBoard();
-  GameBoard.gamesPlayed += 1;
-  console.log('Game started');
-};
+// const initializeGame = () => {
+//   GameBoard.resetBoard();
+//   if (GameBoard.gamesPlayed === 0) {
+//     round.setPlayerNames();
+//   }
+//   round.resetPlayerArrays();
+//   round.moveCount = 0;
+//   displayController.renderBoard();
+//   GameBoard.gamesPlayed += 1;
+//   console.log('Game started');
+// };
 
-displayController.btnListners();
+// displayController.btnListners();
