+++ index.html
@@ -26,6 +26,6 @@
       <div class="block block8"></div>
       <div class="block block9"></div>
     </div>
-    <script type="module" src="./assets/js/dom.js"></script>
+    <script type="module" src="./assets/js/main.js"></script>
   </body>
 </html>
